requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
bcrypt==4.0.1
psycopg2-binary
python-dotenv
supabase
requests
google-genai
Pillow==11.3.0
